[util] Extract reusable focus toggle
