+++ src/automation/company-work.tsx
@@ -1,10 +1,4 @@
-import type {
-    CompanyName,
-    CompanyPositionInfo,
-    Player,
-    NS,
-    UserInterfaceTheme,
-} from 'netscript';
+import type { CompanyName, CompanyPositionInfo, Player, NS } from 'netscript';
 import { MEM_TAG_FLAGS } from 'services/client/memory_tag';
 import {
     KARMA_HEIGHT,
@@ -14,54 +8,7 @@ import {
 
 import { CONFIG } from 'automation/config';
 
-class Toggle {
-    ns: NS;
-    value: boolean;
-
-    constructor(ns: NS, value: boolean) {
-        this.ns = ns;
-        this.value = value;
-    }
-
-    toggle() {
-        this.value = !this.value;
-        this.ns.singularity.setFocus(this.value);
-    }
-}
-
-interface FocusProps {
-    ns: NS;
-    focus: Toggle;
-}
-
-const buttonClass =
-    'MuiButtonBase-root MuiButton-root MuiButton-text MuiButton-textPrimary MuiButton-sizeMedium MuiButton-textSizeMedium MuiButton-root MuiButton-text MuiButton-textPrimary MuiButton-sizeMedium MuiButton-textSizeMedium css-u8jh2y';
-
-function FocusToggle({ ns, focus }: FocusProps) {
-    const [theme, setTheme] = React.useState(
-        ns.ui.getTheme() as UserInterfaceTheme,
-    );
-
-    React.useEffect(() => {
-        const id = globalThis.setInterval(() => {
-            setTheme(ns.ui.getTheme());
-        }, 200);
-
-        return () => {
-            globalThis.clearInterval(id);
-        };
-    }, [ns]);
-
-    return (
-        <button
-            className={buttonClass}
-            style={{ color: theme.successlight }}
-            onClick={() => focus.toggle()}
-        >
-            Toggle Focus
-        </button>
-    );
-}
+import { Toggle, FocusToggle } from 'util/focus';
 
 export async function main(ns: NS) {
     const flags = ns.flags([
