+++ src/util/focus.tsx
@@ -1,0 +1,58 @@
+import type { NS, UserInterfaceTheme } from 'netscript';
+
+/** Toggle focus mode for work actions. */
+export class Toggle {
+    ns: NS;
+    value: boolean;
+
+    constructor(ns: NS, value: boolean) {
+        this.ns = ns;
+        this.value = value;
+    }
+
+    toggle() {
+        this.value = !this.value;
+        this.ns.singularity.setFocus(this.value);
+    }
+}
+
+export interface FocusProps {
+    ns: NS;
+    focus: Toggle;
+}
+
+const buttonClass =
+    'MuiButtonBase-root MuiButton-root MuiButton-text MuiButton-textPrimary MuiButton-sizeMedium MuiButton-textSizeMedium MuiButton-root MuiButton-text MuiButton-textPrimary MuiButton-sizeMedium MuiButton-textSizeMedium css-u8jh2y';
+
+/**
+ * Render a button for toggling focus mode.
+ *
+ * @param ns - Netscript API instance
+ * @param focus - `Toggle` helper controlling state
+ * @returns The focus toggle button
+ */
+export function FocusToggle({ ns, focus }: FocusProps) {
+    const [theme, setTheme] = React.useState(
+        ns.ui.getTheme() as UserInterfaceTheme,
+    );
+
+    React.useEffect(() => {
+        const id = globalThis.setInterval(() => {
+            setTheme(ns.ui.getTheme());
+        }, 200);
+
+        return () => {
+            globalThis.clearInterval(id);
+        };
+    }, [ns]);
+
+    return (
+        <button
+            className={buttonClass}
+            style={{ color: theme.successlight }}
+            onClick={() => focus.toggle()}
+        >
+            Toggle Focus
+        </button>
+    );
+}
